backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-decouple
gunicorn==21.2.0
djangorestframework-simplejwt==5.5.1
dj-rest-auth==7.0.1
django-allauth==64.2.1
requests>=2.31.0  # django-allauthの依存パッケージ